Fix styling issues
Now we match the doc and the selector better, and
the selector must indeed match the tail of the docs.

This means that we don't need to artificially map
a selector to a single element, and e.g. "li" cam
be styled simply if they are embedded into each other.

Closes #221.