Only .auto_close progress bars can be current
I.e. if .auto_close = FALSE, then you need to specify
the id of the progress bar to update and terminate it.

This makes everything much cleaner.

Currently, if the current progress bar changes, then
the previous one is terminated. But this could change
and we could throw an error.

The expectation is that if you need to have more than
two _active_ progress bars in the same frame, then
you would have to use .auto_close = FALSE.

Another alternative would be to introduce a new flag
for this and keep .auto_close just for auto-closing.
Because sometimes (often?) you want auto-closing,
even if you need multiple progress bars from the same
frame.